refactor(cli): extract analyzer confidence capping helpers
