+++ tailtriage-cli/src/analyze/mod.rs
@@ -2,6 +2,7 @@ use std::collections::{BTreeMap, HashMap};
 
 use serde::{Serialize, Serializer};
 
+mod confidence;
 mod evidence;
 
 pub use evidence::{EvidenceQuality, EvidenceQualityLevel, SignalCoverageStatus};
@@ -360,7 +361,7 @@ fn analyze_run_internal(run: &Run) -> Report {
     let warnings = analysis_warnings(run, &suspects);
     let evidence_quality = evidence::evidence_quality(run);
 
-    apply_evidence_aware_confidence_caps(&mut suspects, run, &evidence_quality);
+    confidence::apply_evidence_aware_confidence_caps(&mut suspects, run, &evidence_quality);
 
     let mut ranked = suspects.into_iter();
     let primary_suspect = ranked.next().unwrap_or_else(|| {
@@ -688,171 +689,6 @@ fn has_material_p95_shift(left: Option<u64>, right: Option<u64>) -> bool {
         return false;
     }
     higher.saturating_mul(2) >= lower.saturating_mul(3)
-}
-
-fn apply_evidence_aware_confidence_caps(
-    suspects: &mut [Suspect],
-    run: &Run,
-    evidence_quality: &EvidenceQuality,
-) {
-    let runtime_snapshots_missing = run.runtime_snapshots.is_empty();
-    let runtime_partial_key_fields = !runtime_snapshots_missing
-        && (run
-            .runtime_snapshots
-            .iter()
-            .all(|snapshot| snapshot.blocking_queue_depth.is_none())
-            || run
-                .runtime_snapshots
-                .iter()
-                .all(|snapshot| snapshot.local_queue_depth.is_none())
-            || run
-                .runtime_snapshots
-                .iter()
-                .all(|snapshot| snapshot.global_queue_depth.is_none()));
-    let ambiguous_cluster = ambiguity_cluster_indices(suspects);
-    for (i, suspect) in suspects.iter_mut().enumerate() {
-        let mut cap = Confidence::High;
-        let mut notes = Vec::new();
-        let is_primary = i == 0;
-        let is_insufficient = suspect.kind == DiagnosisKind::InsufficientEvidence;
-        if !is_insufficient && evidence_quality.quality == EvidenceQualityLevel::Weak {
-            cap = cap.min(Confidence::Medium);
-        }
-        if !is_insufficient && run.requests.is_empty() {
-            cap = Confidence::Low;
-            notes.push("Low completed-request count caps confidence.".to_string());
-        } else if run.requests.len() < LOW_COMPLETED_REQUEST_THRESHOLD {
-            if !is_insufficient {
-                cap = cap.min(Confidence::Medium);
-            }
-            if is_primary {
-                notes.push("Low completed-request count caps confidence.".to_string());
-            }
-        }
-        if run.truncation.dropped_requests > 0 && !is_insufficient {
-            cap = cap.min(Confidence::Medium);
-            notes.push(
-                "Capture truncation caps confidence because dropped evidence may affect ranking."
-                    .to_string(),
-            );
-        }
-        apply_family_evidence_caps(
-            &suspect.kind,
-            run,
-            runtime_snapshots_missing,
-            runtime_partial_key_fields,
-            &mut cap,
-            &mut notes,
-        );
-        let ambiguity_capped = ambiguous_cluster.contains(&i) && !is_insufficient;
-        if ambiguity_capped {
-            cap = cap.min(Confidence::Medium);
-            notes.push(
-                "Top suspects are close in score; confidence is capped by ambiguity.".to_string(),
-            );
-        }
-        let original = suspect.confidence;
-        suspect.confidence = original.min(cap);
-        let cap_changed_bucket = suspect.confidence != original;
-        if cap_changed_bucket || ambiguity_capped {
-            notes.sort();
-            notes.dedup();
-            suspect.confidence_notes = notes;
-        } else {
-            suspect.confidence_notes.clear();
-        }
-    }
-}
-
-fn apply_family_evidence_caps(
-    kind: &DiagnosisKind,
-    run: &Run,
-    runtime_snapshots_missing: bool,
-    runtime_partial_key_fields: bool,
-    cap: &mut Confidence,
-    notes: &mut Vec<String>,
-) {
-    match kind {
-        DiagnosisKind::ApplicationQueueSaturation => {
-            if run.truncation.dropped_queues > 0 {
-                *cap = (*cap).min(Confidence::Medium);
-                notes.push(
-                    "Capture truncation caps confidence because dropped evidence may affect ranking."
-                        .to_string(),
-                );
-            }
-            if run.queues.is_empty() {
-                *cap = (*cap).min(Confidence::Medium);
-                notes.push(
-                    "Missing queue instrumentation limits queue-saturation confidence.".to_string(),
-                );
-            }
-        }
-        DiagnosisKind::DownstreamStageDominates => {
-            if run.truncation.dropped_stages > 0 {
-                *cap = (*cap).min(Confidence::Medium);
-                notes.push(
-                    "Capture truncation caps confidence because dropped evidence may affect ranking."
-                        .to_string(),
-                );
-            }
-            if run.stages.is_empty() {
-                *cap = (*cap).min(Confidence::Medium);
-                notes.push(
-                    "Missing stage instrumentation limits downstream-stage confidence.".to_string(),
-                );
-            }
-        }
-        DiagnosisKind::BlockingPoolPressure | DiagnosisKind::ExecutorPressureSuspected => {
-            if run.truncation.dropped_runtime_snapshots > 0 {
-                *cap = (*cap).min(Confidence::Medium);
-                notes.push(
-                    "Capture truncation caps confidence because dropped evidence may affect ranking."
-                        .to_string(),
-                );
-            }
-            if runtime_snapshots_missing {
-                *cap = (*cap).min(Confidence::Medium);
-                notes.push(
-                    "Missing runtime snapshots limit executor/blocking confidence.".to_string(),
-                );
-            } else if runtime_partial_key_fields {
-                *cap = (*cap).min(Confidence::Medium);
-                notes.push(
-                    "Runtime snapshots are partial; missing runtime queue-depth fields limit executor/blocking confidence.".to_string(),
-                );
-            }
-        }
-        DiagnosisKind::InsufficientEvidence => {}
-    }
-}
-
-fn ambiguity_cluster_indices(suspects: &[Suspect]) -> Vec<usize> {
-    let mut ranked = suspects
-        .iter()
-        .enumerate()
-        .filter(|(_, s)| s.kind != DiagnosisKind::InsufficientEvidence)
-        .collect::<Vec<_>>();
-    ranked.sort_by_key(|(_, s)| std::cmp::Reverse(s.score));
-    let Some((_, top)) = ranked.first() else {
-        return Vec::new();
-    };
-    if top.score < AMBIGUITY_MIN_SCORE_THRESHOLD {
-        return Vec::new();
-    }
-    let cluster = ranked
-        .iter()
-        .take_while(|(_, s)| {
-            s.score >= AMBIGUITY_MIN_SCORE_THRESHOLD
-                && top.score.abs_diff(s.score) <= AMBIGUITY_SCORE_GAP_THRESHOLD
-        })
-        .map(|(idx, _)| *idx)
-        .collect::<Vec<_>>();
-    if cluster.len() >= 2 {
-        cluster
-    } else {
-        Vec::new()
-    }
 }
 
 fn clamp_score(value: u64) -> u8 {
@@ -1420,12 +1256,11 @@ mod tests {
     };
 
     use crate::analyze::{
-        analyze_run, analyze_run_internal, apply_evidence_aware_confidence_caps,
-        apply_temporal_overlap_attribution_warning, evidence, render_text, Confidence,
-        DiagnosisKind, EvidenceQuality, EvidenceQualityLevel, InflightTrend, Report,
-        SignalCoverageStatus, Suspect, ROUTE_DIVERGENCE_WARNING, ROUTE_RUNTIME_ATTRIBUTION_WARNING,
-        TEMPORAL_OVERLAP_ATTRIBUTION_WARNING, TEMPORAL_P95_SHIFT_WARNING,
-        TEMPORAL_SUSPECT_SHIFT_WARNING,
+        analyze_run, analyze_run_internal, apply_temporal_overlap_attribution_warning, evidence,
+        render_text, Confidence, DiagnosisKind, EvidenceQuality, EvidenceQualityLevel,
+        InflightTrend, Report, SignalCoverageStatus, Suspect, ROUTE_DIVERGENCE_WARNING,
+        ROUTE_RUNTIME_ATTRIBUTION_WARNING, TEMPORAL_OVERLAP_ATTRIBUTION_WARNING,
+        TEMPORAL_P95_SHIFT_WARNING, TEMPORAL_SUSPECT_SHIFT_WARNING,
     };
 
     fn test_run() -> Run {
@@ -2314,7 +2149,7 @@ mod tests {
             vec![],
             vec![],
         )];
-        apply_evidence_aware_confidence_caps(&mut suspects, &run, &eq);
+        super::confidence::apply_evidence_aware_confidence_caps(&mut suspects, &run, &eq);
         assert!(suspects[0]
             .confidence_notes
             .iter()
@@ -2370,7 +2205,7 @@ mod tests {
             vec![],
             vec![],
         )];
-        apply_evidence_aware_confidence_caps(&mut suspects, &run, &eq);
+        super::confidence::apply_evidence_aware_confidence_caps(&mut suspects, &run, &eq);
         assert!(suspects[0]
             .confidence_notes
             .iter()
@@ -2398,7 +2233,7 @@ mod tests {
             vec![],
             vec![],
         )];
-        apply_evidence_aware_confidence_caps(&mut suspects, &run, &eq);
+        super::confidence::apply_evidence_aware_confidence_caps(&mut suspects, &run, &eq);
         assert_eq!(suspects[0].confidence, Confidence::Medium);
         assert!(suspects[0]
             .confidence_notes
@@ -2422,7 +2257,7 @@ mod tests {
             vec![],
             vec![],
         )];
-        apply_evidence_aware_confidence_caps(&mut suspects, &run, &eq);
+        super::confidence::apply_evidence_aware_confidence_caps(&mut suspects, &run, &eq);
         assert!(suspects[0]
             .confidence_notes
             .iter()
@@ -2442,7 +2277,7 @@ mod tests {
         ];
         let run = test_run();
         let eq = evidence::evidence_quality(&run);
-        apply_evidence_aware_confidence_caps(&mut suspects, &run, &eq);
+        super::confidence::apply_evidence_aware_confidence_caps(&mut suspects, &run, &eq);
         assert_eq!(suspects[0].confidence, Confidence::Medium);
         assert_eq!(suspects[1].confidence, Confidence::Medium);
         assert!(suspects[0]
@@ -2468,7 +2303,7 @@ mod tests {
         ];
         let run = test_run();
         let eq = evidence::evidence_quality(&run);
-        apply_evidence_aware_confidence_caps(&mut suspects, &run, &eq);
+        super::confidence::apply_evidence_aware_confidence_caps(&mut suspects, &run, &eq);
 
         assert_eq!(suspects[0].score, 100);
         assert_eq!(suspects[1].score, 100);
@@ -2521,7 +2356,7 @@ mod tests {
         ];
         suspects[0].confidence = Confidence::High;
         let eq = evidence::evidence_quality(&run);
-        apply_evidence_aware_confidence_caps(&mut suspects, &run, &eq);
+        super::confidence::apply_evidence_aware_confidence_caps(&mut suspects, &run, &eq);
         assert_eq!(suspects[0].confidence, Confidence::High);
         assert!(suspects[0].confidence_notes.is_empty());
     }
